Protected boards route for PUT et DELETE, and automatic linking of userId

+++ src/features/boards/boards/boards.controller.js
@@ -2,10 +2,11 @@ import { parse } from "dotenv";
 import * as boardsService from "./boards.service.js";
 import mongoose from "mongoose";
 
+import { isUserBoardOwner, isUserAdmin } from "./boards.utils.js";
+
 async function getAllBoards(req, res, next) {
   try {
-
-    // validate userId if provided in the query parameters 
+    // validate userId if provided in the query parameters
     let userId = undefined;
 
     let limit = parseInt(req.query.limit) || 2;
@@ -19,16 +20,17 @@ async function getAllBoards(req, res, next) {
       }
     }
 
-
     res.json(await boardsService.getAllBoards(userId, { limit, cursor }));
   } catch (err) {
     next(err);
   }
 }
 
-
 async function createBoard(req, res, next) {
   try {
+    const userId = req.user.id;
+    req.body.userId = userId;
+
     res.json(await boardsService.createBoard(req.body));
   } catch (err) {
     next(err);
@@ -46,7 +48,15 @@ async function getBoardById(req, res, next) {
 
 async function updateBoard(req, res, next) {
   try {
-    console.log("Updating board with ID:", req.params.id, "with data:", req.body);
+    const user = req.user;
+
+    const isOwner = await isUserBoardOwner(req.params.id, user);
+    if (!isOwner) {
+      const error = new Error("Unauthorized: Only the board owner can update this board.");
+      error.status = 403;
+      throw error;
+    }
+
     res.json(await boardsService.updateBoard(req.params.id, req.body));
   } catch (err) {
     next(err);
@@ -55,6 +65,21 @@ async function updateBoard(req, res, next) {
 
 async function deleteBoard(req, res, next) {
   try {
+    const userId = req.user.id;
+    const userRole = req.user.role;
+
+    const isOwner = await isUserBoardOwner(req.params.id, userId);
+    const isAdmin = await isUserAdmin(userRole);
+
+    //if user is not owner, he has to be admin to delete the board
+    if (!isOwner && !isAdmin) {
+      const error = new Error(
+        "Unauthorized: Only the board owner or an admin can delete this board."
+      );
+      error.status = 403;
+      throw error;
+    }
+
     res.json(await boardsService.deleteBoard(req.params.id));
   } catch (err) {
     next(err);
@@ -69,4 +94,11 @@ async function getCommentsByBoardId(req, res, next) {
   }
 }
 
-export {getAllBoards, createBoard, getBoardById, updateBoard, deleteBoard, getCommentsByBoardId};
+export {
+  getAllBoards,
+  createBoard,
+  getBoardById,
+  updateBoard,
+  deleteBoard,
+  getCommentsByBoardId,
+};
